requirements.txt
Django==3.0.6
django-countries==6.1.2
django-extensions==2.2.9
djangorestframework==3.11.0
psycopg2==2.7.5
requests==2.23.0

# coding standards
flake8==3.8.1
pylint==2.5.2
pylint-django==2.0.15